llava/requirements.txt
langchain
langchain-community
Pillow
opencv-python
